refactor(brute): move Finding and KeyMatch to brute/types.go
Output value types belong in their own file, not alongside orchestration
logic. Also tighten BruteResult field assertions in result_test.go:
add inline comments explaining intent and assert KeyMatch.CVE round-trips.
