+++ brute/result_test.go
@@ -3,6 +3,7 @@ package brute
 import "testing"
 
 func TestBruteResultCarriesFinding(t *testing.T) {
+	// shows Finding can coexist with non-auth result
 	r := &BruteResult{
 		ConnectionSuccess: true,
 		Finding: &Finding{
@@ -17,6 +18,7 @@ func TestBruteResultCarriesFinding(t *testing.T) {
 }
 
 func TestBruteResultCarriesKeyMatch(t *testing.T) {
+	// success path: KeyMatch + AuthSuccess together
 	r := &BruteResult{
 		AuthSuccess:       true,
 		ConnectionSuccess: true,
@@ -28,5 +30,8 @@ func TestBruteResultCarriesKeyMatch(t *testing.T) {
 	}
 	if r.KeyMatch == nil || r.KeyMatch.Vendor != "Vagrant" {
 		t.Fatalf("KeyMatch not carried on BruteResult")
+	}
+	if r.KeyMatch.CVE != "CVE-2015-1338" {
+		t.Fatalf("KeyMatch.CVE = %q, want CVE-2015-1338", r.KeyMatch.CVE)
 	}
 }
